export multiple LOD

+++ src/test/java/org/xpen/dunia2/fileformat/xbg/HalfFloatUtilTest.java
@@ -7,9 +7,9 @@ public class HalfFloatUtilTest {
 	
     @Test
     public void testToFloat() {
-    	Assert.assertEquals(0f, HalfFloatUtil.toFloat(0), 0.01f);
-    	Assert.assertEquals(-40480f, HalfFloatUtil.toFloat(-1807), 0.01f);
-    	Assert.assertEquals(0f, HalfFloatUtil.toFloat(2610), 0.01f);
+    	Assert.assertEquals(0f, HalfFloatUtil.toFloat((short)0), 0.01f);
+    	Assert.assertEquals(-40480f, HalfFloatUtil.toFloat((short)-1807), 0.01f);
+    	Assert.assertEquals(0f, HalfFloatUtil.toFloat((short)2610), 0.01f);
     }
 
 }
